Add userId in all elements

+++ Recipes/src/Application/Command/Book/RemoveBook.php
@@ -2,8 +2,10 @@
 
 namespace Recipes\Application\Command\Book;
 
+use LIN3S\SharedKernel\Exception\Exception;
 use Recipes\Domain\Model\Book\BookId;
 use Recipes\Domain\Model\Book\BookRepository;
+use Recipes\Domain\Model\User\UserId;
 
 class RemoveBook
 {
@@ -16,6 +18,10 @@ class RemoveBook
 
     public function __invoke(RemoveBookCommand $command)
     {
+        $book = $this->repository->bookOfId(BookId::generate($command->id()));
+        if(!$book->owner()->equals(UserId::generate($command->userId()))) {
+            throw new Exception('Invalid userId, is not your book');
+        }
         $this->repository->remove(BookId::generate($command->id()));
     }
 }
